put main project page in ProjectMain component for chunking
