+++ client/src/pages/Projects.js
@@ -1,4 +1,5 @@
 import React, { useState, useEffect } from "react";
+import ProjectMain from "../components/ProjectMain";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { Container, Row, Col, Button, ListGroup } from "react-bootstrap";
 import { faGithub } from "@fortawesome/free-brands-svg-icons";
@@ -16,56 +17,11 @@ function Projects() {
     const { data } = await API.getProjects();
     console.log(data);
   };
-  const goToProject = () => {
-    console.log("click");
-    setSpotlight(false);
-  };
 
   return (
     <div>
       {spotlight ? (
-        <Container className="projects__cont">
-          <h2 className="projects__h2">Projects</h2>
-          <Row>
-            <Col md={6}>
-              <figure className="projects__media">
-                <img
-                  className="projects__img"
-                  src="/img/trowel-snapshot-2.png"
-                  alt="snapshot of project website"
-                  onClick={goToProject}
-                />
-              </figure>
-            </Col>
-            <Col className="projects__summary">
-              <h2 className="projects__summary-title">Trowel</h2>
-              <p className="projects__summary--blurb">
-                An App for landscapers to look up valuable plant information and
-                organize gardens by their clients.
-              </p>
-              <ul className="projects__summary--list-group">
-                <li className="projects__summary--list-items">
-                  {" "}
-                  ➢ Logo Design
-                </li>
-                <li className="projects__summary--list-items">
-                  {" "}
-                  ➢ Front End Development
-                </li>
-                <li className="projects__summary--list-items">
-                  {" "}
-                  ➢ Back End Development
-                </li>
-              </ul>
-              <Button
-                className="projects__view-project-btn"
-                onClick={goToProject}
-              >
-                View Project
-              </Button>
-            </Col>
-          </Row>
-        </Container>
+        <ProjectMain />
       ) : (
         <Container className="projects__spotlight-cont">
           <Row>
